Refactors at event and club modules
1. Fixed an issue at the add new event page where the site returns a 403 Forbidden error on form data submsision.
2. Updated the delete event function to fix an issue where deleting the event does not clear the event images from storage.
3. Moved remaining business logic from event and club controllers to services.
4. Minor edits to view-all-clubs and view-all-events.
5. Fixed an issue where the event bookmarks page could not be loaded.
6. Minor edits to edit club img and edit event img views.
