+++ resources/views/events-finder/bookmarks.blade.php
@@ -72,7 +72,9 @@
                         <div class="row pb-3 px-md-3 px-sm-0">
                             @foreach ($bookmarks as $bookmark)
                                 <div class="col-xl-2 col-lg-3 col-md-4 col-6 mb-3 px-2">
-                                    <x-event-card :event="$bookmark->event"/>
+                                    <x-event-card
+                                        :event="$bookmark->event"
+                                        :intersectionarray="$intersectionArray"/>
                                 </div>
                             @endforeach
                         </div>
@@ -81,7 +83,9 @@
                     <div id="list-view" class="row list-view ms-2 {{ $searchViewPreference == 2 ? '' : 'd-none' }} mt-0">
                         @foreach ($bookmarks as $bookmark)
                             <div class="row pb-3">
-                                <x-event-list-item :event="$bookmark->event"/>
+                                <x-event-list-item
+                                    :event="$bookmark->event"
+                                    :intersectionarray="$intersectionArray"/>
                             </div>
                         @endforeach
                     </div>
